edit groups tab gets headers (carl); use sampo's SSE find_peaks code; fix build for find_peaks on x86; don't duplicate sources when embedding; use Glib::ustring for all source-related strings; fixup plugin UI automation buttons
git-svn-id: svn://localhost/ardour2/trunk@1595 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/playlist.cc
@@ -1508,7 +1508,7 @@ Playlist::bump_name_once (string name)
 	} else {
 		int isnumber = 1;
 		const char *last_element = name.c_str() + period + 1;
-		for (int i = 0; i < strlen(last_element); i++) {
+		for (size_t i = 0; i < strlen(last_element); i++) {
 			if (!isdigit(last_element[i])) {
 				isnumber = 0;
 				break;
@@ -1525,7 +1525,7 @@ Playlist::bump_name_once (string name)
 		} else {
 			char buf[32];
 
-			snprintf (buf, sizeof(buf), "%d", version+1);
+			snprintf (buf, sizeof(buf), "%ld", version+1);
 		
 			newname  = name.substr (0, period+1);
 			newname += buf;
